show custom messages for erros

+++ lista-react/src/containers/Auth/Login.js
@@ -6,11 +6,16 @@ import { connect, dispatch } from 'react-redux'
 
 import Button from '../../components/Button/Button'
 import Input from '../../components/Input/Input'
+import Popup from '../../components/Popup/Popup'
+
+import withErrorHandler from '../../hoc/withErrorHandler/withErrorHandler'
 
 import classes from './Auth.module.sass'
 
 const Login = (props) => {
     const [update, setUpdate] = useState(false)
+    const [popupDisplay, setPDisplay] = useState(false)
+    const [popupText, setPText] = useState(null)
 
     useEffect(() => {
         console.log('wasap')
@@ -20,35 +25,59 @@ const Login = (props) => {
         //document.querySelector('#loginContainer').requestFullscreen()
     })
 
-    const handleSubmit = () => {
+    const handleSubmit = async () => {
         const username = document.getElementById('username').value
         const password = document.getElementById('password').value
-        axios.post('login', {
-            username: username,
-            password: password
-        })
-            .then(res => {
-                document.exitFullscreen()
-                console.log(res.data)
+        try {
+            let res= await axios.post('login', {
+                username: username,
+                password: password
+            })
+            if (!res.data.token){                  //me he estresado y ya no lo voy a hacer con errores
+                console.log(res.data.msg)
+                setPText(res.data.msg)
+                setPDisplay(true)
+                console.log(popupDisplay)
+                setTimeout(() => {setPDisplay(false)}, 5000)
+            } else {
+
+                if(document.fullscreenElement !== null){
+                    document.exitFullscreen()
+                }
+                console.log(res, res.response, res.error)
                 props.setToken(res.data.token)
                 localStorage.clear()
                 localStorage.setItem('token', JSON.stringify(res.data.token))
                 props.history.push('/listas')
-            })
-            .catch(err => {
-                console.log(err)
-                //axios.post(`http://192.168.1.12:8080/auth/login`, {username: username, password: password})
-                    //.then(res => {
-                        //console.log(res.data)
-                        //props.setToken(res.data.token)
-                        //localStorage.clear()
-                        //localStorage.setItem('token', JSON.stringify(res.data.token))
-                        //props.history.push('/listas')
-                        //})
-                    //.catch(err => console.log(err))
-            })
+
+            }
+        } catch (error){
+            console.log(error.response)
+            console.log('emosido enganiado')
+            setPText(error.response ? error.response.data.msg : error.message)
+            setPDisplay(true)
+            console.log(popupDisplay)
+            setTimeout(() => {setPDisplay(false)}, 5000)
+        }
+            //.then(res => {
+
+                //if(document.fullscreenElement !== null){
+                    //document.exitFullscreen()
+                //}
+                //console.log(res, res.response, res.error)
+                //props.setToken(res.data.token)
+                //localStorage.clear()
+                //localStorage.setItem('token', JSON.stringify(res.data.token))
+                //props.history.push('/listas')
+            //})
+            //.catch(err => {
+                //console.log('[hola]', err, err.response, err.data, err.msg)
+            //})
     }
+    console.log(popupDisplay)
     return(
+        <>
+        <Popup display={popupDisplay} type="error">{popupText}</Popup>
         <div id="loginContainer" className={classes.bigContainer}>
             <form className={classes.form}>
                 <h1 className={classes.h1}>log in</h1>
@@ -62,6 +91,7 @@ const Login = (props) => {
                 </div>
             </form>
         </div>
+        </>
     )
 }
 
@@ -75,4 +105,4 @@ const mapActions = (dispatch) => {
     }
 }
 
-export default connect(null, mapActions)(Login)
+export default connect(null, mapActions)/*(withErrorHandler*/(Login/*, axios)*/)
